Add dataset for testing VBA script
This dataset is a testing dataset with sample data. It is much shorter in length compared to the dataset future VBA scripts will be applied too. However, it is formatted similarly, thus results should correlate. This dataset will allow for performant testing. The end result will be a script that can be applied to this dataset, and those like it, for instant analysis.